requirements.txt
streamlit
langchain==0.3.14
langchain-community==0.3.14
langchain-groq
langchain-huggingface
sentence-transformers
langchain-text-splitters==0.3.5
langchain-core==0.3.33
chromadb
pypdf
unstructured
rank_bm25
python-dotenv
